Session id format added,Send to Client added, added test cases for Ransha and input sub-protocols

+++ mpc/src/common/rbc/mod.rs
@@ -30,7 +30,7 @@ pub enum RbcError {
     InvalidThreshold(usize, usize),
 
     #[error("Session {0} already ended")]
-    SessionEnded(usize),
+    SessionEnded(u64),
 
     #[error("Unknown Bracha message type: {0}")]
     UnknownMsgType(String),
